Fix: app-admin 신고 관리 community 응답 형태 불일치 수정

+++ services/app-admin-service/src/main/java/com/gb/appadmin/domain/report/dto/response/MemberReportItemResponse.java
@@ -10,7 +10,7 @@ public record MemberReportItemResponse(
         String authorPublicId,
         String targetType,
         String category,
-        int reportCount,
+        long reportCount,
         String status,
         String lastReportedAt
 ) {}
